pyproject.toml
[project]
name = "pubTools"
version = "0.0.1"
description = "A collection of LaTeX/Gerby tools to help publish the diSimplex documents"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "plastex>=3.1",
    "lpilPlasTeX @ git+https://github.com/diSimplex/plastex-lpil-plugin.git",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
dsplastex = "pubTools.plastex:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.20.3",
    "pytest-mock>=3.10.0",
]
